chore: Meter, ValueRange, Time and Chrono now use a double type internally instead of a float

+++ Source/CkCore/Public/CkCore/Meter/CkMeter_Data.h
@@ -18,7 +18,7 @@ public:
 
 private:
     UPROPERTY(VisibleAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = true))
-    float _CurrentValue = 0.0f;
+    double _CurrentValue = 0.0;
 
 public:
     CK_PROPERTY_GET(_CurrentValue);
@@ -39,7 +39,7 @@ public:
 
 private:
     UPROPERTY(VisibleAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = true))
-    float _Size = 0.0f;
+    double _Size = 0.0;
 
 public:
     CK_PROPERTY_GET(_Size);
@@ -60,7 +60,7 @@ public:
 
 private:
     UPROPERTY(VisibleAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = true))
-    float _AmountRemaining = 0.0f;
+    double _AmountRemaining = 0.0;
 
 public:
     CK_PROPERTY_GET(_AmountRemaining);
@@ -81,7 +81,7 @@ public:
 
 private:
     UPROPERTY(VisibleAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = true))
-    float _AmountUsed = 0.0f;
+    double _AmountUsed = 0.0;
 
 public:
     CK_PROPERTY_GET(_AmountUsed);
@@ -101,8 +101,8 @@ public:
     CK_GENERATED_BODY(FCk_Meter_Replenish);
 
 private:
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, meta = (AllowPrivateAccess = true, ClampMin = 0.0f, UIMin = 0.0f))
-    float _ReplenishAmount = 0.0f;
+    UPROPERTY(EditAnywhere, BlueprintReadWrite, meta = (AllowPrivateAccess = true, ClampMin = 0.0, UIMin = 0.0))
+    double _ReplenishAmount = 0.0;
 
 public:
     CK_PROPERTY_GET(_ReplenishAmount);
@@ -122,8 +122,8 @@ public:
     CK_GENERATED_BODY(FCk_Meter_Consume);
 
 private:
-    UPROPERTY(EditAnywhere, BlueprintReadWrite, meta = (AllowPrivateAccess = true, ClampMin = 0.0f, UIMin = 0.0f))
-    float _ConsumeAmount = 0.0f;
+    UPROPERTY(EditAnywhere, BlueprintReadWrite, meta = (AllowPrivateAccess = true, ClampMin = 0.0, UIMin = 0.0))
+    double _ConsumeAmount = 0.0;
 
 public:
     CK_PROPERTY_GET(_ConsumeAmount);
@@ -144,10 +144,10 @@ public:
 
 private:
     UPROPERTY(EditAnywhere, BlueprintReadWrite, meta = (AllowPrivateAccess = true))
-    float _MinCapacity = 0.0f;
+    double _MinCapacity = 0.0;
 
     UPROPERTY(EditAnywhere, BlueprintReadWrite, meta = (AllowPrivateAccess = true))
-    float _MaxCapacity = 0.0f;
+    double _MaxCapacity = 0.0;
 
 public:
     CK_PROPERTY(_MinCapacity);
@@ -172,7 +172,7 @@ private:
     FCk_Meter_Capacity _Capacity;
 
     UPROPERTY(EditAnywhere, BlueprintReadWrite, meta = (AllowPrivateAccess = true))
-    FCk_FloatRange_0to1 _StartingPercentage = FCk_FloatRange_0to1{0.0f};
+    FCk_FloatRange_0to1 _StartingPercentage = FCk_FloatRange_0to1{0.0};
 
 public:
     CK_PROPERTY_GET(_Capacity);
